Add Silent Payments address types
Add SilentPaymentsDestination for v0 to v30; Bip352 reserves v31
for a backwards incompatible change. v1 to v30 addresses are handled
as specified in Bip352. The extra_data will be ignored during sending.

Valid v0 and v1 addresses were added to key_io_valid.json for testing,
and invalid v0 and a v31 address were added to key_io_invalid.json.

SilentPaymentsDestination must not be formed from invalid public keys;
the operations that must be performed to send to a SilentPaymentsDestination,
require valid public keys. Hence, the SilentPaymentsDestination struct takes
the following precautions to ensure that such invalid states are not possible:
- SilentPaymentsDestination instances can only be constructed from a public static factory
  that returns nullopt if the parameters are invalid.
- The public keys are private and const references can be retrieved using member functions.